Fix the OCR findings on M1 and M2
- test_arena_pool called irq_enable_all(), which lives in the optional
  irq module the test does not link (a -FromSources build fails); it
  uses __builtin_sti() now. Its irq_save() results are read into
  variables before CHECK_EQ, which evaluates its arguments again when
  it reports a failure.
- test_math compares rsqrt over every kind of float with the others,
  uses the float_bits function as the oracle, and checks the
  float_bits/float_from_bits macros against their functions per value.
- test_memory checks that the _fast names are the standard routines.
- sys_sp() is cast to unsigned int like the function it replaces.
- Comments: fma stays a call because Ceres-C has no __builtin_fma;
  clz/ctz of 0 are 32 by Ceres-C's own definition of its builtins; the
  abs instruction gives INT_MIN back for INT_MIN and nothing folds it.

Dismissed: the byte references in test_memory and test_string_fast
cannot turn into the routines under test - Ceres-C lowers such loops
to routines it emits itself (__cc_memcpy, __cc_strlen, ...), not to
the library's.

+++ include/ceres/sys.h
@@ -22,7 +22,7 @@ void sys_set_features(unsigned int features);
 void sys_panic(const char* msg) __attribute__((noreturn));   // print "panic: <msg>" on the terminal and shut down
 
 unsigned int sys_sp(void);           // the stack pointer, as seen by this call (asm/sys.casm) ...
-#define sys_sp() __builtin_stack_pointer()   // ... read in place, with no call
+#define sys_sp() ((unsigned int)__builtin_stack_pointer())   // ... read in place, with no call
 unsigned int sys_heap_start(void);   // the linker's __heap_start: first free byte above the image
 unsigned int sys_stack_free(void);   // bytes between the top of the heap and sp
